updating image paths
Reason previous version did not work was because I had put a "../" in front of the file path/Name
While this is a very common syntax, and Visual Studio recognizes this syntax and was able to display image just fine, running the file from a native web browser directly from windows explorer, did not recognize this syntax. - Requesting reevaluation of the grade. The synatx was not wrong for the IDE I was using